refactor(rust): extract builders/state from deliverability (≤300 LOC)
