+++ src/bin/email_api_dir/admin_ops/diag_handlers/mod.rs
@@ -1,9 +1,11 @@
 #![allow(unused_imports, dead_code)]
 
 pub mod deliverability;
+pub mod deliverability_builders;
 pub mod observability;
 pub mod security;
 
 pub use deliverability::*;
+pub use deliverability_builders::*;
 pub use observability::*;
 pub use security::*;
